perf+cleanup: byte-identical low-severity optimizations + dead-code
Batch 1 of the low-severity sweep (verified by full suite, 2059 passed):
- 37 byte-identical perf micro-opts across processing/neuropixels/recording/
  modeling/analyses/visualizations: hoist loop-invariants, cache repeated reads/
  parses/globs, reuse already-computed arrays, drop redundant copies, in-place
  ops, list.index()->dict, dedup I/O.
- dead-code: removed unreachable except in StringTuple.convert.

Two flagged items were reverted as not actually safe: usv_count/usv_rate
find_bout_epochs keys (a test consumes them as output contract) and the sglx
augment_legacy_meta pre-parsed-meta param (passed dict lacked imroTbl).

+++ src/usv_playpen/processing/export_yolo_dataset.py
@@ -264,7 +264,7 @@ class YOLODatasetExporter:
                     for row in np.flatnonzero(durations > 0):
                         split = "val" if position in val_positions else "train"
                         position += 1
-                        spec = specs[row].astype(np.float32)
+                        spec = specs[row]
                         duration = int(durations[row])
                         spec_id = f"{session_id}_{int(row)}"
